Modify Him.java to provide farewell message when user exits the program overall
